Don't use .comm directive
This is doing something completely else! It (globally) allocates space in bss segment so it was totally pointless to put it in .data and even .text (XBRA was not working at all).

Replaced also in bss sections as we lose the option to align stuff plus .ds.X makes it more readable in general. Kept only in mintaudio as those are the only variables referenced globally.