atari:video:xbios: use vdi to store/restore palette
There's no reason why this should be done on XBIOS level when the
application is, after all, always returning to GEM. This avoids
implementation of various hardware-specific routines and just rely on
the driver to do its job properly.

Of course, setting a new palette is still done on XBIOS level.