atari:audio: delete all code and start anew
The main reason is to reduce the crazy number of drivers available. They
are mutually exclusive most of the time so there's no point in offering
them to the user separately. Also the framework didn't work well on the
FireBee and in general had issues hard to fix and track.

TODO: check for usound.h in ./configure